Thiết lập lại cấu trúc DB và các lớp dữ liệu

+++ backend/Web_Trang_Suc_BackEnd/Models/Entities/CartItem.cs
@@ -8,13 +8,13 @@ namespace web_Trang_suc_BE.Models.Entities
     {
         [Key]
         [Column("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Column("cart_id")]
-        public int CartId { get; set; }
+        public long CartId { get; set; }
 
         [Column("product_variant_id")]
-        public int ProductVariantId { get; set; }
+        public long ProductVariantId { get; set; }
 
         [Column("quantity")]
         public int Quantity { get; set; } = 1;
